+++ camera/models.py
@@ -2,9 +2,14 @@
 
 from __future__ import unicode_literals
 
-from django.db import models
-from django.utils.text import slugify
 import os
+
+from django.db import models
+from django.db.models.signals import post_save
+from django.dispatch import receiver
+from django.utils.text import slugify
+
+from .utils import detect_faces_on_image, detect_eyes_on_image
 
 def upload_images_path(instance, filename):
     """
@@ -18,13 +23,21 @@ def upload_images_path(instance, filename):
         ext=ext
     ))
 
+
 class Image(models.Model):
     """
     Image model - to be uploaded in Django-Admin.
     """
 
+    TYPE_CHOICES = (
+        (0, 'Face recognition'),
+        (1, 'Eyes recognition'),
+    )
+
     title = models.CharField('Image title', max_length=255)
     image = models.ImageField('Image file', upload_to=upload_images_path, blank=True, null=True)
+    date_created = models.DateTimeField(auto_now_add=True, editable=False)
+    type = models.BooleanField('Type of processing', help_text='Choose a proper algorithm.', choices=TYPE_CHOICES, default=0)
 
     def __str__(self):
         return self.title
@@ -32,3 +45,13 @@ class Image(models.Model):
     class Meta:
         verbose_name = 'Image'
         verbose_name_plural = 'Images'
+
+
+@receiver(post_save, sender=Image, dispatch_uid='detect_faces')
+def detect_faces(sender, instance, **kwargs):
+    if kwargs.get('created', False):
+        if instance.type == 0:
+            detect_faces_on_image(instance)
+        else:
+            detect_eyes_on_image(instance)
+        instance.save()