+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 * Created README.md, CHANGELOG.md files;
 * Created 'env' requirements.txt file;
 * Created .gitignore;
+* Created app `camera` with method for detecting faces;